[TvShow] Move tv show model files to tvShows directory
